refactor(e2e): share ziti diagnostics helper
