+++ suites/go-core/suite.yaml
@@ -66,6 +66,11 @@ run: |
   fi
   set -e
 
+  if [ "$status" -ne 0 ]; then
+    E2E_ZITI_DIAGNOSTICS_DUMP=1 \
+      go test -count=1 -timeout 1m -tags "${tag_args}" -run '^TestZitiDiagnosticsDump$' -v ./tests/... || true
+  fi
+
   go run github.com/jstemmer/go-junit-report/v2@v2.1.0 -parser gojson < "${tmp_json}" > junit.xml
   rm -f "${tmp_json}"
